+++ src/Recyclarr.Cli/Pipelines/Generic/IConfigPipelinePhase.cs
@@ -3,5 +3,5 @@ namespace Recyclarr.Cli.Pipelines.Generic;
 public interface IConfigPipelinePhase<in TContext>
     where TContext : IPipelineContext
 {
-    Task Execute(TContext context);
+    Task Execute(TContext context, CancellationToken ct);
 }