+++ src/Recyclarr.Compatibility/Radarr/RadarrCapabilities.cs
@@ -4,10 +4,7 @@ namespace Recyclarr.Compatibility.Radarr;
 //
 // May get used one day; keep the parameter around so that calling
 // code does not need to be changed later.
-public record RadarrCapabilities(Version? Version)
+public record RadarrCapabilities(Version Version)
 {
-    public RadarrCapabilities()
-        : this((Version?) null)
-    {
-    }
+    public bool QualityDefinitionLimitsIncreased => Version >= new Version(5, 9, 0, 9049);
 }