+++ src/Recyclarr.TrashGuide/QualitySize/IQualityItemLimits.cs
@@ -1,7 +1,0 @@
-namespace Recyclarr.TrashGuide.QualitySize;
-
-public interface IQualityItemLimits
-{
-    decimal MaxLimit { get; }
-    decimal PreferredLimit { get; }
-}