+++ tests/Recyclarr.Tests.TestLibrary/NewQualitySize.cs
@@ -7,6 +7,8 @@ public static class NewQualitySize
     public static QualityItemWithLimits WithLimits(string quality, decimal min, decimal max, decimal preferred)
     {
         var item = new QualityItem(quality, min, max, preferred);
-        return new QualityItemWithLimits(item, new TestQualityItemLimits());
+        return new QualityItemWithLimits(item, new QualityItemLimits(
+            TestQualityItemLimits.MaxUnlimitedThreshold,
+            TestQualityItemLimits.PreferredUnlimitedThreshold));
     }
 }